GameManager.cs
Removed SpawnManager.cs and replaced with GameManager.cs
